tests/tesseract/requirements.txt
ABBYY==0.3
alembic==0.9.7
certifi==2017.11.5
chardet==3.0.4
click==6.7
cssselect==1.0.3
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
numpy==1.14.0
opencv-python==3.4.0.12
pandas==0.22.0
pdf2image==0.1.7
pdfminer==20140328
pdfquery==0.4.3
pdftabextract==0.3.0
pdftables-api==1.0.0
Pillow==5.0.0
PyPDF2==1.26.0
pyquery==1.4.0
pytesseract==0.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
requests==2.18.4
roman==2.0.0
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.1
tabula-py==1.0.0
urllib3==1.22
visitor==0.1.3
Wand==0.4.4
Werkzeug==0.14.1
WTForms==2.1
